Updated code to pass unit test after re-organizing into sub-modules based on use-case. Added debug category
